--- __mocks__/db/ExtesnionRequest.ts
@@ -1,80 +1,0 @@
-import { ExtensionRequest } from '@/components/ExtensionRequest/ExtensionStatusModal';
-
-export const now = Date.now();
-
-export const mockExtensionRequests: ExtensionRequest[] = [
-    {
-        id: '1',
-        reason: 'Need more time',
-        newEndsOn: now + 86400000,
-        title: 'Fix bugs',
-        taskId: '123',
-        oldEndsOn: now,
-        status: 'APPROVED',
-        requestNumber: 1,
-        timestamp: now,
-        assignee: 'john',
-        assigneeId: 'user-1',
-        reviewedBy: 'admin',
-        reviewedAt: Math.floor(now / 1000),
-    },
-    {
-        id: '2',
-        reason: 'Additional requirements',
-        newEndsOn: now + 172800000,
-        title: 'Add features',
-        taskId: '123',
-        oldEndsOn: now - 86400000,
-        status: 'PENDING',
-        requestNumber: 2,
-        timestamp: now - 43200000,
-        assignee: 'john',
-        assigneeId: 'user-1',
-    },
-];
-
-export const mockExtensionRequestDetails: ExtensionRequest[] = [
-    {
-        id: '1',
-        reason: 'Need more time',
-        newEndsOn: now + 86400000,
-        title: 'Fix bugs',
-        taskId: '123',
-        oldEndsOn: now,
-        status: 'APPROVED',
-        requestNumber: 1,
-        timestamp: now,
-        assignee: 'john',
-        assigneeId: 'user-1',
-        reviewedBy: 'admin',
-        reviewedAt: 1619090400,
-    },
-    {
-        id: '2',
-        reason: 'Additional requirements',
-        newEndsOn: now + 172800000,
-        title: 'Add features',
-        taskId: '123',
-        oldEndsOn: now - 86400000,
-        status: 'DENIED',
-        requestNumber: 2,
-        timestamp: now - 43200000,
-        assignee: 'john',
-        assigneeId: 'user-1',
-        reviewedBy: 'manager',
-        reviewedAt: 1714611746,
-    },
-    {
-        id: '3',
-        reason: 'Task complexity increased',
-        newEndsOn: now + 172800000,
-        title: 'Implement new API',
-        taskId: '123',
-        oldEndsOn: now - 86400000,
-        status: 'PENDING',
-        requestNumber: 3,
-        timestamp: now - 43200000,
-        assignee: 'john',
-        assigneeId: 'user-1',
-    },
-];
